--- a/src/NSubstitute/Core/Arguments/ArgumentFormatter.cs
+++ b/src/NSubstitute/Core/Arguments/ArgumentFormatter.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Linq;
-using System.Reflection;
-
-namespace NSubstitute.Core.Arguments
+﻿namespace NSubstitute.Core.Arguments
 {
     public class ArgumentFormatter : IArgumentFormatter
     {
@@ -17,18 +13,8 @@ namespace NSubstitute.Core.Arguments
             if (arg == null) return "<null>";
             if (arg is string) return string.Format("\"{0}\"", arg);
             var standardToString = arg.ToString();
-            if (standardToString == arg.GetType().ToString()) return FormatType(arg.GetType());
+            if (standardToString == arg.GetType().ToString()) return arg.GetType().GetNonMangledTypeName();
             return standardToString;
        }
-
-        private string FormatType(Type type)
-        {
-            var typeName = type.Name;
-            if (!type.GetTypeInfo().IsGenericType) return typeName;
-
-            typeName = typeName.Substring(0, typeName.IndexOf('`'));
-            var genericArgTypes = type.GetGenericArguments().Select(x => FormatType(x));
-            return string.Format("{0}<{1}>", typeName, string.Join(", ", genericArgTypes.ToArray()));
-        }
     }
 }