--- a/src/NSubstitute/Proxies/CastleDynamicProxy/CastleDynamicProxyFactory.cs
+++ b/src/NSubstitute/Proxies/CastleDynamicProxy/CastleDynamicProxyFactory.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
@@ -25,14 +26,22 @@ namespace NSubstitute.Proxies.CastleDynamicProxy
         {
             VerifyClassHasNotBeenPassedAsAnAdditionalInterface(additionalInterfaces);
 
-            var interceptor = new CastleForwardingInterceptor(
+            var proxyIdInterceptor = new ProxyIdInterceptor(typeToProxy);
+            var forwardingInterceptor = new CastleForwardingInterceptor(
                 new CastleInvocationMapper(
                     new CallFactory(),
                     _argSpecificationDequeue),
                 callRouter);
+
             var proxyGenerationOptions = GetOptionsToMixinCallRouter(callRouter);
-            var proxy = CreateProxyUsingCastleProxyGenerator(typeToProxy, additionalInterfaces, constructorArguments, interceptor, proxyGenerationOptions);
-            interceptor.StartIntercepting();
+            var proxy = CreateProxyUsingCastleProxyGenerator(
+                typeToProxy,
+                additionalInterfaces,
+                constructorArguments,
+                new IInterceptor[] {proxyIdInterceptor, forwardingInterceptor},
+                proxyGenerationOptions);
+
+            forwardingInterceptor.StartIntercepting();
             return proxy;
         }
 
@@ -47,15 +56,28 @@ namespace NSubstitute.Proxies.CastleDynamicProxy
 
         private object CreateProxyUsingCastleProxyGenerator(Type typeToProxy, Type[] additionalInterfaces,
                                                             object[] constructorArguments,
-                                                            IInterceptor interceptor,
+                                                            IInterceptor[] interceptors,
                                                             ProxyGenerationOptions proxyGenerationOptions)
         {
             if (typeToProxy.GetTypeInfo().IsInterface)
             {
                 VerifyNoConstructorArgumentsGivenForInterface(constructorArguments);
-                return _proxyGenerator.CreateInterfaceProxyWithoutTarget(typeToProxy, additionalInterfaces, proxyGenerationOptions, interceptor);
+
+                var interfaces = new List<Type> {typeToProxy};
+                if (additionalInterfaces != null)
+                    interfaces.AddRange(additionalInterfaces);
+
+                // We need to create a proxy for the object type, so we can intercept the ToString() method.
+                // Therefore, we put the desired primary interface to the secondary list.
+                typeToProxy = typeof(object);
+                additionalInterfaces = interfaces.ToArray();
             }
-            return _proxyGenerator.CreateClassProxy(typeToProxy, additionalInterfaces, proxyGenerationOptions, constructorArguments, interceptor);
+
+            return _proxyGenerator.CreateClassProxy(typeToProxy,
+                additionalInterfaces,
+                proxyGenerationOptions,
+                constructorArguments,
+                interceptors);
         }
 
         private ProxyGenerationOptions GetOptionsToMixinCallRouter(ICallRouter callRouter)
@@ -65,10 +87,30 @@ namespace NSubstitute.Proxies.CastleDynamicProxy
             return options;
         }
 
+        private static void VerifyNoConstructorArgumentsGivenForInterface(object[] constructorArguments)
+        {
+            if (constructorArguments != null && constructorArguments.Length > 0)
+            {
+                throw new SubstituteException("Can not provide constructor arguments when substituting for an interface.");
+            }
+        }
+
+        private static void VerifyClassHasNotBeenPassedAsAnAdditionalInterface(Type[] additionalInterfaces)
+        {
+            if (additionalInterfaces != null && additionalInterfaces.Any(x => x.GetTypeInfo().IsClass))
+            {
+                throw new SubstituteException("Can not substitute for multiple classes. To substitute for multiple types only one type can be a concrete class; other types can only be interfaces.");
+            }
+        }
+
         private class AllMethodsExceptCallRouterCallsHook : AllMethodsHook
         {
             public override bool ShouldInterceptMethod(Type type, MethodInfo methodInfo)
             {
+                // Always intercept object.ToString() as we would like to return proxy id as a result.
+                if (ProxyIdInterceptor.IsDefaultToStringMethod(methodInfo))
+                    return true;
+
                 return IsNotCallRouterMethod(methodInfo)
                     && IsNotBaseObjectMethod(methodInfo)
                     && base.ShouldInterceptMethod(type, methodInfo);
@@ -81,23 +123,7 @@ namespace NSubstitute.Proxies.CastleDynamicProxy
 
             private static bool IsNotBaseObjectMethod(MethodInfo methodInfo)
             {
-                return methodInfo.GetBaseDefinition().DeclaringType != typeof (object);
-            }
-        }
-
-        private void VerifyNoConstructorArgumentsGivenForInterface(object[] constructorArguments)
-        {
-            if (constructorArguments != null && constructorArguments.Length > 0)
-            {
-                throw new SubstituteException("Can not provide constructor arguments when substituting for an interface.");
-            }
-        }
-
-        private void VerifyClassHasNotBeenPassedAsAnAdditionalInterface(Type[] additionalInterfaces)
-        {
-            if (additionalInterfaces != null && additionalInterfaces.Any(x => x.GetTypeInfo().IsClass))
-            {
-                throw new SubstituteException("Can not substitute for multiple classes. To substitute for multiple types only one type can be a concrete class; other types can only be interfaces.");
+                return methodInfo.GetBaseDefinition().DeclaringType != typeof(object);
             }
         }
     }