Closes #44

In this PR I've implemented the proxy id feature. which allows to recognize the particular instance of the proxy. The ID looks like following:
```
Substitute.PrimaryProxyType#42
```

Where `PrimaryProxyType` - the short type name of the primary proxy type, `#42` - auto-incremented field (global counter, not per type).

To achieve the desired behavior I've changed the following:
1. Always make a substitute for the class. If interfaces are specified only, we substitute for the `System.Object`. That is required to intercept the `ToString()` method, otherwise Castle doesn't hook it.
2. Add the `ProxyIdInterceptor` to return proxy id when `ToString()` is called. Given that our interceptor participates earlier than forwarding one, calls to the `ToString()` are transparent for the NSubstutite core.
3. We do not intercept `ToString()` if type has its own implementation.

I've tested the code and it works fine in VS debugger - proxy ID is shown for the debugger view:
![image](https://user-images.githubusercontent.com/1074182/37346189-594500c4-26d7-11e8-95e9-5f572abbc516.png)

Due to some reasons Rider doesn't want to evaluate `ToString()`, however we can still see the proxy ID:
![image](https://user-images.githubusercontent.com/1074182/37346279-96923d7a-26d7-11e8-97b0-67b81d40be76.png)

Hope one day they will improve their evaluator and it will invoke the complex `ToString()` method as well.

Of course, everything works fine with assertions, as they invoke `ToString()` method:
```c#
        [Test]
        public void CollectionAssertion()
        {
            var array = new[] {Substitute.For<IInterface>(), Substitute.For<IInterface>()};
            var item = Substitute.For<IInterface>();
            
            Assert.That(array, Contains.Item(item));
        }
```
and error:
```
  Expected: some item equal to <NSubstitute.Proxy.IInterface#3>
  But was:  < <NSubstitute.Proxy.IInterface#1>, <NSubstitute.Proxy.IInterface#2> >
```

-----
I don't like a bit that now we are always forced to create proxy for the class. However, that should not affect the performance too much and the benefit seems to worth it.

@alexandrnikitin @dtchepak Please take a look and let me know your opinion. Thanks 🌹