@zvirja 2018-03-14
UPDATE. I've changed ID to make it shorter. Now it looks like following:
```
Substitute.PrimaryProxyType#42
```

It should allow to save space in the different error messages and looks nicer.