@zvirja 2018-03-22
@dtchepak  See my replies in the comments as well.

> I was initially expecting this to be implemented via the CallRouter mixed-in via ProxyGenerationOptions. If this logic could go in CallRouter it might give us a good hook to support explicitly tagging substitutes later on? (like this example, where we might want to say sub1.Name("good"); sub2.Name("bad") or similar.)

I thought that, but later realized that it's a bad idea. If we decide to move call through the `CallRouter`, it will modify the last active router on dispatch. It might have very bad consequences for the whole NSubsitute functionality, because this method is sometimes called implicitly by e.g.:
- `string.Format()`
- string concatentation
- debugger, when you hover your mouse.

I want `ToString()` invocation to be:
- completely transparent for all our business logic, so we don't worry about those invocations at all.
- very fast, so there are less chances that debugger will refuse to evaluate it.

The first point might be achieved by adding the extra conditions inside the `CallRouter.Route()` and `CastleInvocationMapper`. But that will force us to always remember about this special case and one day we'll overlook something 😉 

The feature with explicit naming will be still possible even with current architecture, as it's possible to pass `ICallRouter` instance to the `ProxyIdInterceptor`. So it could check e.g. `router.SubstituteName` property and return it if it's not null. It's only one of the possible ways - believe we have the better ones.

Given that I prefer to have a separate interceptor as the best guarantee of isolation. I believe that overhead is negligible.

> Have you done any quick measures on perf differences with this extra interception? Might be worth running a quick comparison and making sure it doesn't add much overhead on construction or interception.

I'll follow up on that a bit later, when I have more free time.