@zvirja 2018-03-23
@dtchepak Here are the performance testing results:

Code
```c#
public class NSubstituteBenchmark
{
    private ClassWithIntMethod _classProxy;
    private IInterfaceWithIntMethod _interfaceProxy;

    public NSubstituteBenchmark()
    {
        this._classProxy = Substitute.For<ClassWithIntMethod>();
        this._interfaceProxy = Substitute.For<IInterfaceWithIntMethod>();
    }

    [Benchmark]
    public ClassWithIntMethod CreateClassProxy() => Substitute.For<ClassWithIntMethod>();

    [Benchmark]
    public IInterfaceWithIntMethod CreateInterfaceProxy() => Substitute.For<IInterfaceWithIntMethod>();

    [Benchmark]
    public string InvokeToStringInterfaceProxy() => _interfaceProxy.ToString();

    [Benchmark]
    public string InvokeToStringClassProxy() => _classProxy.ToString();

    [Benchmark]
    public int InvokeRegularMethodInterfaceProxy() => _interfaceProxy.GetInt();

    [Benchmark]
    public int InvokeRegularMethodClassProxy() => _classProxy.GetInt();

    public interface IInterfaceWithIntMethod
    {
        int GetInt();
    }

    public abstract class ClassWithIntMethod
    {
        public abstract int GetInt();
    }
}

```

Environment:
```
BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 2 [1703, Creators Update] (10.0.15063.786)
Intel Core i7-4770 CPU 3.40GHz (Haswell), 1 CPU, 8 logical cores and 4 physical cores
Frequency=3320311 Hz, Resolution=301.1766 ns, Timer=TSC
  [Host]     : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2115.0
  DefaultJob : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2115.0
```

Before:
```
                            Method |        Mean |      Error |      StdDev |
---------------------------------- |------------:|-----------:|------------:|
                  CreateClassProxy | 5,011.48 ns | 95.3044 ns | 105.9305 ns |
              CreateInterfaceProxy | 5,124.30 ns | 26.9081 ns |  23.8533 ns |
      InvokeToStringInterfaceProxy |    10.95 ns |  0.0948 ns |   0.0887 ns |
          InvokeToStringClassProxy |    10.77 ns |  0.0822 ns |   0.0642 ns |
 InvokeRegularMethodInterfaceProxy | 4,636.77 ns | 77.1161 ns |  64.3955 ns |
     InvokeRegularMethodClassProxy | 6,981.80 ns | 63.3611 ns |  52.9094 ns |
```

After:
```
                            Method |        Mean |      Error |     StdDev |
---------------------------------- |------------:|-----------:|-----------:|
                  CreateClassProxy | 5,134.43 ns | 58.4588 ns | 54.6824 ns |
              CreateInterfaceProxy | 5,539.99 ns | 71.3262 ns | 63.2288 ns |
      InvokeToStringInterfaceProxy |    48.91 ns |  1.1124 ns |  1.8890 ns |
          InvokeToStringClassProxy |    47.88 ns |  0.1797 ns |  0.1593 ns |
 InvokeRegularMethodInterfaceProxy | 4,728.07 ns | 59.1856 ns | 49.4227 ns |
     InvokeRegularMethodClassProxy | 6,997.70 ns | 30.7895 ns | 22.2628 ns |
```

As you can see, there is almost no difference in performance and overhead from extra interceptor is negligible.